refactor: :art: prepare for merging
Resolve conflicts with branch dev before merging. Created an additional ClientComm.java file for Client connection simulation (old App.java from feature/authentication).
