+++ client/src/main/java/dev/nextchat/client/ClientComm.java
@@ -1,0 +1,54 @@
+/*
+ * This source file was generated by the Gradle 'init' task
+ */
+package dev.nextchat.client;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.PrintWriter;
+import java.net.Socket;
+
+public class ClientComm {
+    public String getGreeting() {
+        return "Hello World!";
+    }
+    
+    public static void main(String[] args) throws IOException {
+        try {
+            Socket s = new Socket("localhost", 1234);
+            System.out.println("Connected to server.");
+
+            BufferedReader in = new BufferedReader(new InputStreamReader(s.getInputStream())); 
+            PrintWriter out = new PrintWriter(s.getOutputStream(), true); 
+            BufferedReader input = new BufferedReader(new InputStreamReader(System.in)); 
+
+            // This is the purpose of test and simulation only 
+            System.out.println("Please type 1 to sign-in or 2 to sign-up: ");
+
+            String choice = input.readLine().trim();
+
+            while ( !(choice.equals("1") || choice.equals("2")) ) {
+                System.out.println("Invalid input. Please type 1 to sign-in or 2 to sign-up: "); 
+                choice = input.readLine().trim();
+            }
+
+            out.println(choice); 
+
+
+            System.out.println("Please type in your username: ");
+            out.println(input.readLine()); 
+
+            System.out.println("Please type in your password: ");
+            out.println(input.readLine());
+
+            String response = in.readLine();
+            System.out.println("Server response: " + response);            
+        } 
+        
+        catch (Exception e) 
+        {
+            e.printStackTrace();
+        }
+    }
+}
