+++ server/build.gradle
@@ -27,6 +27,10 @@ dependencies {
     // Spring Boot with MongoDB
     implementation 'org.springframework.boot:spring-boot-starter-data-mongodb'
 
+    implementation 'com.googlecode.json-simple:json-simple:1.1.1'
+    
+    implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.15.0'
+    
     implementation 'org.springframework.boot:spring-boot-starter-security'
 
     implementation 'jakarta.persistence:jakarta.persistence-api:3.1.0'
