+++ server/src/main/java/dev/nextchat/server/App.java
@@ -12,6 +12,16 @@ import java.net.ServerSocket;
 import java.net.Socket;
 import java.security.NoSuchAlgorithmException;
 
+// import org.springframework.beans.factory.annotation.Autowired;
+// import org.springframework.boot.CommandLineRunner;
+// import org.springframework.boot.SpringApplication;
+// import org.springframework.boot.autoconfigure.SpringBootApplication;
+// import org.springframework.context.annotation.ComponentScan;
+// import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
+
+// @SpringBootApplication
+// @ComponentScan(basePackages = {"dev.netchat.server.utils"})
+// @EnableMongoRepositories(basePackages = "com.chatapp.repository")
 public class App {
     public String getGreeting() {
         return "Hello World!";
